sm6150-common: Generate non-hlos image mount points for A/B  builds
On builds where A/B support is enabled the system image acts as the
rootfs. In this case we can no longer create the non-hlos image mount
points at run time as we used to. We now create them as part of the
build itself. As for the non A/B cases it does not hurt having them
created at build time and simplifies selinux rules.

Change-Id: I23a6ccc29d1abd30937fcacaf4a7071c2314b70e
Signed-off-by: IceWinterBot506 <mydadisbrave3@gmail.com>